+++ bin/server.dart
@@ -1,10 +1,14 @@
+library server;
+
 import "dart:io";
 
 import 'package:logging_handlers/server_logging_handlers.dart';
 import 'package:logging/logging.dart';
 
-import 'package:restful/restful.dart';
-import 'shoping_application.dart';
+import 'package:restful/rest_server.dart';
+
+part 'service.dart';
+part 'customer.dart';
 
 /**
  * Main entry point.
@@ -15,7 +19,7 @@ main() {
 	int port = 8080;
 	
 	// Logger
-	Logger.root.onRecord.listen(new PrintHandler()); // default PrintHandler
+	Logger.root.onRecord.listen(new PrintHandler());
 	Logger.root.level = Level.FINE;
 	Logger logger = new Logger("Server");
   
@@ -23,45 +27,16 @@ main() {
   HttpServer.bind(host, port).then((server) {
   	logger.info("Server listens [${host}:${port}]");
   	logger.info("To stop server press CTRL + C");
-    // Create new rest runtime
-    RestRuntime rest = new RestRuntime();
-    // register all applications
-    rest.register([new ShoppingApllication()]);
+    // Create new REST Server
+  	RestServer rest = new RestServer([new CustomerService()]);
     // Start listenning
     server.listen((HttpRequest request) {
-      logger.fine("Processing ${request.uri.path}");
-      // Process REST services 
-      if (!rest.service(request)) {
-        // Process other type of requests or resources 
-        final Uri uri = request.uri;
-        if (!uri.isAbsolute) {
-          final String stringPath = uri.path == '/' ? 'index.html' : uri.path;
-          final File file = new File(stringPath);
-          file.exists()
-            ..then((bool found) {
-            if (found) {
-              file.openRead()
-                .pipe(request.response)
-                .catchError((e) { 
-                  logger.severe("File i/o error", e);
-                });
-            } else {
-              // File not found
-              logger.warning("File not found ${stringPath}");
-              _error404(request);
-            }
-          })
-          ..catchError((error){
-            // File doesn't exists
-            logger.warning("File doesn't exists ${stringPath}");
-            _error404(request);
-          });
-        } else {
-          // URI incorrect
-          logger.warning("URI incorrect");
-          _error404(request);
-        }
-      };
+      logger.fine("Request to ${request.uri.path}");
+      try {
+        rest.service(request);
+      } on Exception catch(e) {
+        _error404(request);
+      }
     });
   });
 }
@@ -70,7 +45,41 @@ main() {
  * Error 404.
  */
 void _error404(HttpRequest request) {
-  request.response.write("<h1>Error</h1><p>404 - File not found</p>");
+  request.response.write("<h1>Error</h1><p>404 - Page not found</p>");
   request.response.statusCode = HttpStatus.NOT_FOUND;
   request.response.close();
 }
+
+/*
+//      if (!rest.service(request)) {
+//        // Process other type of requests or resources 
+//        final Uri uri = request.uri;
+//        if (!uri.isAbsolute) {
+//          final String stringPath = uri.path == '/' ? 'index.html' : uri.path;
+//          final File file = new File(stringPath);
+//          file.exists()
+//            ..then((bool found) {
+//            if (found) {
+//              file.openRead()
+//                .pipe(request.response)
+//                .catchError((e) { 
+//                  logger.severe("File i/o error", e);
+//                });
+//            } else {
+//              // File not found
+//              logger.warning("File not found ${stringPath}");
+//              _error404(request);
+//            }
+//          })
+//          ..catchError((error){
+//            // File doesn't exists
+//            logger.warning("File doesn't exists ${stringPath}");
+//            _error404(request);
+//          });
+//        } else {
+//          // URI incorrect
+//          logger.warning("URI incorrect");
+//          _error404(request);
+//        }
+//      };
+*/